+++ src/plugin/SettingsTab.tsx
@@ -42,7 +42,12 @@ function Calendars({
             style={{ marginRight: '4px' }}
             onClick={() => {
               if (!confirm('Remove this calendar?')) return
-              _.pull(plugin.settings.calendars, calendar)
+              console.log('updated version')
+
+              plugin.settings.calendars = _.pull(
+                [...plugin.settings.calendars],
+                calendar
+              )
               plugin.saveSettings()
               updateCalendars()
             }}